+++ src/components/overlays/ActivateServiceSubscription/index.tsx
@@ -317,7 +317,9 @@ export default function ActivateserviceSubscription({
                   loadIndicator="Loading ..."
                   loading
                   size="medium"
-                  onButtonClick={() => {}}
+                  onButtonClick={() => {
+                    // do nothing
+                  }}
                   sx={{ marginLeft: '10px' }}
                 />
               ) : (