+++ src/components/overlays/AddMultipleUser/index.tsx
@@ -443,7 +443,9 @@ export default function AddMultipleUser() {
               loadIndicator={t('global.actions.loading')}
               loading
               size="medium"
-              onButtonClick={() => {}}
+              onButtonClick={() => {
+                // do nothing
+              }}
               sx={{ marginLeft: '10px', textTransform: 'none' }}
             />
           ) : (