+++ src/components/overlays/EditUsecase/index.tsx
@@ -138,7 +138,9 @@ export default function EditUsecase({
             loadIndicator="Loading ..."
             loading
             size="medium"
-            onButtonClick={() => {}}
+            onButtonClick={() => {
+              // do nothing
+            }}
             sx={{ marginLeft: '10px' }}
           />
         ) : (