+++ src/components/overlays/InviteForm/index.tsx
@@ -151,7 +151,9 @@ export const InviteForm = ({
               loadIndicator={t('global.actions.loading')}
               loading
               size="medium"
-              onButtonClick={() => {}}
+              onButtonClick={() => {
+                // do nothing
+              }}
               sx={{ marginLeft: '10px' }}
             />
           ) : (